2025年秋季-硕士研究生/第三组：cvc5自底向上——从基础到应用/requirements.txt
# cvc5 SMT Solver Project Dependencies

# Core dependency - cvc5 SMT solver
cvc5>=1.2.0

# Data manipulation and numerical computing
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: For better visualization
seaborn>=0.12.0

# Optional: For performance profiling
pytest-benchmark>=4.0.0

# Code quality (optional, for development)
black>=23.0.0
flake8>=6.0.0
pysmt>=0.9.6

# LLM Integration (for natural language constraint solver)
openai>=1.0.0
python-dotenv>=1.0.0
